+++ types/input/createInputValueParser.d.ts
@@ -1,0 +1,1 @@
+export { createInputValueParser, createInputValueParser as default } from "../input.js";