+++ types/input/parseInputValue.d.ts
@@ -1,0 +1,1 @@
+export { parseInputValue, parseInputValue as default } from "../input.js";